Change call to Discord ID Snowflake for Windows

+++ source/Engine/Extensions/Discord.cpp
@@ -393,7 +393,7 @@ int Discord::User::Update() {
 	}
 
 	CurrentUser.IDSnowflake = user.id;
-	snprintf(CurrentUser.ID, sizeof CurrentUser.ID, "%" PRId64, CurrentUser.IDSnowflake);
+	snprintf(CurrentUser.ID, sizeof(CurrentUser.ID), "%lld", (long long)CurrentUser.IDSnowflake);
 	StringUtils::Copy(CurrentUser.Username, user.username, sizeof(CurrentUser.Username));
 	CurrentUser.IsBot = user.bot;
